Fixed projectile damage
