+++ GMDFinal/GMDProject/Assets/Scenes/Test-Devlin.unity
@@ -349,11 +349,62 @@ MonoBehaviour:
   m_LightCookieSize: {x: 1, y: 1}
   m_LightCookieOffset: {x: 0, y: 0}
   m_SoftShadowQuality: 0
+--- !u!1 &1740341957 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 2912724059962043203, guid: 50f6c423499da694994e5e12fec8c17c, type: 3}
+  m_PrefabInstance: {fileID: 6497569005437864085}
+  m_PrefabAsset: {fileID: 0}
 --- !u!4 &1740341963 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 1213718066348686921, guid: 50f6c423499da694994e5e12fec8c17c, type: 3}
   m_PrefabInstance: {fileID: 6497569005437864085}
   m_PrefabAsset: {fileID: 0}
+--- !u!61 &1740341964
+BoxCollider2D:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1740341957}
+  m_Enabled: 1
+  serializedVersion: 3
+  m_Density: 1
+  m_Material: {fileID: 0}
+  m_IncludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ExcludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_LayerOverridePriority: 0
+  m_ForceSendLayers:
+    serializedVersion: 2
+    m_Bits: 4294967295
+  m_ForceReceiveLayers:
+    serializedVersion: 2
+    m_Bits: 4294967295
+  m_ContactCaptureLayers:
+    serializedVersion: 2
+    m_Bits: 4294967295
+  m_CallbackLayers:
+    serializedVersion: 2
+    m_Bits: 4294967295
+  m_IsTrigger: 1
+  m_UsedByEffector: 0
+  m_CompositeOperation: 0
+  m_CompositeOrder: 0
+  m_Offset: {x: 0, y: 0}
+  m_SpriteTilingProperty:
+    border: {x: 0, y: 0, z: 0, w: 0}
+    pivot: {x: 0.5, y: 0.5}
+    oldSize: {x: 1, y: 1}
+    newSize: {x: 1, y: 1}
+    adaptiveTilingThreshold: 0.5
+    drawMode: 0
+    adaptiveTiling: 0
+  m_AutoTiling: 0
+  m_Size: {x: 1, y: 1}
+  m_EdgeRadius: 0
 --- !u!1001 &97548259135551776
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -475,10 +526,14 @@ PrefabInstance:
       propertyPath: fireRate
       value: 0.25
       objectReference: {fileID: 0}
-    m_RemovedComponents: []
+    m_RemovedComponents:
+    - {fileID: 660835580667956658, guid: 50f6c423499da694994e5e12fec8c17c, type: 3}
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
-    m_AddedComponents: []
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: 2912724059962043203, guid: 50f6c423499da694994e5e12fec8c17c, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 1740341964}
   m_SourcePrefab: {fileID: 100100000, guid: 50f6c423499da694994e5e12fec8c17c, type: 3}
 --- !u!1001 &7852013822427152108
 PrefabInstance:
@@ -494,11 +549,15 @@ PrefabInstance:
       objectReference: {fileID: 9100000, guid: eb984caded128d746941e9595d4c3b3b, type: 2}
     - target: {fileID: 640184677313334876, guid: a29dae6dd119b1246aa29e3456629d87, type: 3}
       propertyPath: speed
-      value: 2.5
+      value: 2
       objectReference: {fileID: 0}
     - target: {fileID: 640184677313334876, guid: a29dae6dd119b1246aa29e3456629d87, type: 3}
       propertyPath: vertical
       value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 640184677313334876, guid: a29dae6dd119b1246aa29e3456629d87, type: 3}
+      propertyPath: maxHealth
+      value: 2
       objectReference: {fileID: 0}
     - target: {fileID: 2243444367387775122, guid: a29dae6dd119b1246aa29e3456629d87, type: 3}
       propertyPath: m_LocalPosition.x
